Make the ui_input crate just about generic input components
